backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apify-client>=1.8.1
vaderSentiment==3.3.2
networkx==3.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
joblib==1.4.2
pymongo==4.8.0
